+++ x-project/community-system/community-common/src/main/java/cn/xin/learn/community/entity/params/integral/PageIntegralParam.java
@@ -23,7 +23,7 @@ public class PageIntegralParam extends PageParam {
     //归属用户
     private String belongUser;
     //归属用户Id
-    private Integer belongUserId;
+    private Long belongUserId;
     //积分类型（0-使用、1-新增）
     private Integer integralType;
