+++ x-project/community-system/community-common/src/main/java/cn/xin/learn/community/entity/po/Integral.java
@@ -23,7 +23,7 @@ public class Integral extends BasePo {
     //归属用户
     private String belongUser;
     //归属用户Id
-    private Integer belongUserId;
+    private Long belongUserId;
     //积分类型（0-使用、1-新增）
     private Integer integralType;
